Run job "now()" after a scheduler restart.
When adding a new account its necessary to restart scheduler after which it waits the default interval time before next run. If there are setup issues in the new account you'd be waiting for a while to resolve and try again. Hence the simple flag to make it start "now()"

15 minutes could save you... :)