Fix for _find_dependent_monitors
Type: Bugfix

Why is this change necessary:
When find_changes is run, it has the ability to update specific tech
types outside of the normal reporter run. In these cases the monitor
must rerun auditors for the updated tech types, plus any other auditors
for other tech types that are dependent on the updated types. It
determines this by checking the support_watcher_indexes and
support_auditor_indexs for each auditor associated with the tech types
updated. The check for the support_watcher_indexes were only being run
for the last auditor because it was indented incorrectly.

This change addresses the need by:
Changing the indent so that it is run for all auditors