Add support for custom account metadata (#408) 📃
* Add support for custom watchers and auditors

Change summary:
  * Fix Error in exception handling
  * Adding audit class to distinguish items created by different
    auditors of the same type
  * Adding custom directories
  * Adding development guidelines documentation
  * Fixing links in documentation
  * Removing duplicate auditors
  * Adding init test for scheduler
  * Fixing DB error in scheduler test
  * Removing unneeded DB insert from test case
  * Race condition with watcher_registry
  * Fixing DB migration conflict
  * Picking up pubspec.lock changes
  * Code style cleanup
  * Update file headers for contribution prep

Change-Id: Id72322f6dbccedc701e9c17dc9a5b8dc26bf30c1

* Add new watchers

List of added watchers:
    * CloudTrail
    * AWSConfig
    * AWSConfigRecorder
    * DirectConnect::Connection
    * DirectConnect::VirtualGateway
    * EC2::EbsSnapshot
    * EC2::EbsVolume
    * EC2::Image
    * EC2::Instance
    * ENI
    * KMS::Grant
    * KMS::Key
    * Lambda
    * RDS::ClusterSnapshot
    * RDS::DBCluster
    * RDS::DBInstace
    * RDS::Snapshot
    * RDS::SubnetGroup
    * Route53
    * Route53Domains
    * TrustedAdvisor
    * VPC::DHCP
    * VPC::Endpoint
    * VPC::FlowLog
    * VPC::NatGateway
    * VPC::NetworkACL
    * VPC::Peering

Additional changes:
  * Move rds[security_group] to rds/ directory.
  * Update vpc/route_table to use boto3 lib.
  * Add tests for tech types supported by moto
  * Initialize name to avoid UnboundLocalError
  * Update RDS watcher ephemeral values
  * Use boto3.session.get_available_regions in select watchers
  * Convert routetable watcher to decorator pattern
  * Convert route53 domains to decorator pattern
  * Handle the case where the aws principle is a string
  * Catching assume role exception in decorator

* Auditor dependency and link support

* Fix for _find_dependent_monitors

Type: Bugfix

Why is this change necessary:
When find_changes is run, it has the ability to update specific tech
types outside of the normal reporter run. In these cases the monitor
must rerun auditors for the updated tech types, plus any other auditors
for other tech types that are dependent on the updated types. It
determines this by checking the support_watcher_indexes and
support_auditor_indexs for each auditor associated with the tech types
updated. The check for the support_watcher_indexes were only being run
for the last auditor because it was indented incorrectly.

This change addresses the need by:
Changing the indent so that it is run for all auditors

* Add framework for custom account types

Type: generic-large-feature

Why is this change necessary?
Feature allowing for custom account types, which can either be
non-AWS accounts or AWS accounts extended with additional attributes.
By default, the new custom attributes are stored in the database but
can be configured to be retrieved from some other source.

This change addresses the need by:
Implementing account_manager framework