[secmonkey] Fix item 'first_seen' query (#446) 👓
Type: generic-bugfix

Why is this change necessary?
The 'First Seen' datetime on the Items page was showing as the Last Modified
time.
Netflix bug report https://github.com/Netflix/security_monkey/issues/377

This change addresses the need by:
The view tried to reorder the item.revisions relationship to ascending, which
did not work because it was decending in the relationship declaration. Added
a seperate ascending query to get the first seen

Potential Side Effects:
Adds additional DB queries to the items list page, but the impact should be
very small because of pagination. The dashboard used a different path.