Fixing two policy diff cases. (#453) 🛁
1) If provided with a string, json.loads() is used, which often means
strings in the extracted structure are of type unicode.  If the same
structure is also provided as a dict, there will often be problems where
a str and a unicode are seen as different objects.  To fix, we simply dump
any datastructures and read them back in with the json lib.

2) The lib was not properly handling null values and was ignoring them entirely
in the output.