Fix account join issue
Introduced (partially) in #872. You can't filter on a table that hasn't
been properly joined.

Also, you can't join the same table several times, so the existing code
would have also broken if account and accounttype were filtered at the
same time.

So, pull the join out, and invoke it when anyone needs it.